Fix [-fpermissive] compilation warnings/errors.

+++ src/mitsubishi2mqtt/config.h
@@ -41,7 +41,7 @@ const PROGMEM  uint8_t redLedPin = 0;
 // Define global variables for network
 const PROGMEM char* hostnamePrefix = "HVAC_";
 const PROGMEM uint32_t WIFI_RETRY_INTERVAL_MS = 300000;
-int wifi_timeout;
+unsigned long wifi_timeout;
 bool wifi_config_exists;
 String hostname = "";
 String ap_ssid;
